requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
apscheduler>=3.10.0
pillow>=10.0.0
openai>=1.20.0
